Лог Фазы 1: DHCP-сервер прибора не поднимался вообще (ESP_ERR_ESP_NETIF_DHCPS_START_FAILED) — base.flags затирал ESP_NETIF_DHCP_SERVER, объект dhcps не создавался; флаги netif исправлены, порядок dhcps_stop->set_ip_info->dhcps_start, ожидание адреса 15->6 с, подсказка 'хост в другой подсети' для раздачи; поворот 270 хранится 16 битами; новый tools/idf_diag_read.py; разбор лога в docs/bugs-2026-09-17.md

+++ idf/main/display.cpp
@@ -48,7 +48,7 @@ static void fb_px(int x, int y, bool black)
 }
 
 /* --------------------------------------------------- поворот кадра (настройка) */
-static uint8_t s_rot_deg;              /* 0 / 90 / 180 / 270 */
+static uint16_t s_rot_deg;             /* 0 / 90 / 180 / 270 (270 в uint8_t не влезает) */
 static uint8_t s_rot_buf[FB_LEN];      /* временный буфер для поворота */
 
 static bool fb_get(const uint8_t *b, int x, int y)
@@ -72,7 +72,7 @@ static void fb_put(uint8_t *b, int x, int y, bool black)
     }
 }
 
-void display_set_rotation(uint8_t deg)
+void display_set_rotation(uint16_t deg)
 {
     if (deg != 90 && deg != 180 && deg != 270) {
         deg = 0;
